Compile to modern targets using browserslist
Browserslist is the easiest way to select a group of browsers for CSS
and JS compilation, although it's less fine grained than feature lists.
Gotta think about this, how to make an argument for one over the other
-- I guess w/ browserslist we're saying we're OK with some users not
being covered.

This is also a first attempt, I need to diff packages to see if it has
done anything.